refactor components and move api calls to services
