+++ src/Components/index.js
@@ -2,6 +2,11 @@ export { HomePage } from "./HomePage/HomePage";
 export { Navbar } from "./Navbar/Navbar";
 export { NavPane } from "./NavPane/NavPane";
 export { Playlists } from "./Playlists/Playlists";
+export { Playlist } from "./Playlists/Playlist";
 export { VideoCard } from "./VideoCard/VideoCard";
 export { Watch } from "./Watch/Watch";
 export { User } from "./User/User";
+export { LikedVideos } from "./LikedVideos/LikedVideos";
+export { PlaylistModal } from "./PlaylistModal/PlaylistModal";
+export { Login } from "./Authentication/Login";
+export { Signup } from "./Authentication/Signup";
